src/udf-optimizer/requirements.txt
# Original test dependencies
python-dotenv
google-generativeai

# Parallelization implementation dependencies
pyyaml

# Optional: For real DeerFlow integration
# langgraph
# langchain-core
# langchain-openai